Confirmation delete. and onclick method for edit

+++ resources/views/todo/show.blade.php
@@ -4,7 +4,7 @@
 @section('content')
     <div class="form-group row">
         <div class="col-xs-2">
-            <h1 id="title">{{ $todo->name }} <img src="/imgs/pencil.png" alt=""></h1>
+            <h1 for="inputdefault" id="title">{{ $todo->name }} <img src="/imgs/pencil.png" alt="pencil" style="cursor: pointer;"></h1>
         </div>
     </div>
 
@@ -16,10 +16,11 @@
         <div class="col-sm-4">
             <div class="card">
                 <div class="card-body">
-    <a href="/task/{{$task->id}}">{{ $task->content }}</a><a href="/delete/task/{{$task->id}}"class="float-right">x</a>
+                    <a href="/task/{{$task->id}}">{{ $task->content }}</a><a href="/delete/task/{{$task->id}}" onclick="return confirm('Are you sure to delete?')" class="float-right">x</a>
                 </div>
             </div>
-        </div>@endforeach
+        </div>
+    @endforeach
 </ul>
 
 <hr>
